Create and process iRODS data requests

Create an iRODS delete request for an object or collection

Create an iRODS delete request for an object or collection

Usage

sodar_delete_request(
  project_uuid,
  path,
  description = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_list(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_cancel(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_reject(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_accept(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_delete_request(
  project_uuid,
  path,
  description = NULL,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_list(
  project_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_cancel(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_reject(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_request_accept(
  request_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

Arguments

project_uuid

A character string containing the UUID of the project.

path

A character string containing the path to the object or collection.

description

A character string containing the description of the request.

config

A list containing the SODAR configuration. If NULL, it tries to get it from the global options.

verbose

A logical value indicating whether to make verbose requests.

return_raw

A logical value indicating whether to return the raw response.

request_uuid

A character string containing the UUID of the request (returned by sodar_delete and sodar_request_list).