Handling existing landing zones

This functions retrieve information about a landing zone for a project, delete existing landing zones etc.

Usage

sodar_lz_retrieve(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_lz_status(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE)
)

sodar_lz_delete(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_lz_move(
  lz_uuid,
  move = TRUE,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

sodar_lz_validate(
  lz_uuid,
  config = NULL,
  verbose = getOption("verbose", default = FALSE),
  return_raw = FALSE
)

Arguments

lz_uuid

A character string containing the UUID of the landing zone.

config

A list containing the SODAR configuration. If NULL, it tries to get it from the global options.

verbose

A logical value indicating whether to make verbose requests.

return_raw

A logical value indicating whether to return the raw response.

move

A logical value indicating whether to move the landing zone after validation.

Value

A list with the details of the landing zone.